+++ src/realloc.c
@@ -1,4 +1,4 @@
-#include "include/realloc.h"
+#include "../include/realloc.h"
 
 // Function to copy data from one block to another
 void copy_block(t_block src, t_block dst)
